requirements.txt
fastapi>=0.70.1
pandas==1.2.4
pydantic==1.8.1
requests==2.25.1
starlette>=0.13.6
#torch==1.8.0+cu111
transformers==4.5.1
gunicorn==20.1.0
uvicorn[standard] >= 0.16.0
anyio==3.4.0
torch-scatter==2.0.6